:recycle: Remove validate.rs
