+++ src/internal.rs
@@ -4,4 +4,3 @@ pub mod test_utils;
 
 pub mod gf2_linalg;
 pub mod utils;
-pub mod validate;
